validation_study/requirements.txt
krippendorff==0.5.1
matplotlib==3.6.1
numpy==1.23.4
pandas==1.5.0
scikit-learn==1.1.2
tabulate==0.9.0
python-docx==0.8.11

